+++ posthog-server/CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Next
 
-## 2.2.0 - 2025-12-01
+- feat: Add ETag support for local evaluation polling to reduce bandwidth when flags haven't changed ([#350](https://github.com/PostHog/posthog-android/pull/350))
 
-- feat: include `evaluated_at` properties in `$feature_flag_called` events ([#321](https://github.com/PostHog/posthog-android/pull/321))
+## 2.1.0 - 2025-12-05
+
+- feat: Include `evaluated_at` properties in `$feature_flag_called` events ([#321](https://github.com/PostHog/posthog-android/pull/321))
+- feat: Add `appendFeatureFlags` optional boolean to `capture` ([#347](https://github.com/PostHog/posthog-android/pull/347))
 
 ## 2.0.1 - 2025-11-24